Add arguments to YML directives
resolves DiUS/java-faker#134 (closed PR)

The idea to use arguments in resolve was closely related to something i wanted to do which is to put the power of
letterify, numberify and regexify into the YML expression language.  This allows for more complicated expressions
that can be sculpted to be more correct without having to write java.

Now you can use something like #{regexify '[45]{2}ABC'} and it will work returning either 4ABC or 5ABC.  This should
help in creating more correct phone numbers, area codes, etc...

Any method exposed on Faker, even with arguments will work.

Even #{bothify '##??','true'} will work.  While all arguments are 'strings' in the YML, they are coerced to the appropriate
type prior to calling the method in Faker, so this expression will result in faker.bothify("##??", true).