Add es-MX and en-ZA yml files.
Fixes DiUS/java-faker#153

The problem was with a fix in de89bf74 .  The fix used a simple toLowerCase
on the directive which didn't adequetly transform it.  This manifested itself
in an unresolved directive value in the en-ZA an es-MX file.  The es-MX file
contained #{PhoneNumber.lada_tres} as a directive and did contain phone_number > lada_tres
but when the FakeValueService attempted to transform the directive it transformed it to
phonenumber.lada_tres erroneously.  The fix in this commit properly converts
PhoneNumber.lada_tres > phone_number.lada_tres which then properly resolves as a simple
reference within the yml file.