Fix an issue with expression resolution order.
The expression resolution was checking the yml file BEFORE the current object.
That's a problem because of something like 'building_number'.  building_number is
implemented in address by calling numerify(resolve('building_number')) but
the resolution order had yml first instead of the object, so calling address.streetAddress() was returning
 ### my street name instead of a random number like '104 my street name'

This commit fixes this issue