Always return valid ISBN Numbers (#172)
* Allow digits to be in expression names like isbn_gs1

* Always return valid ISBN numbers

Implement valid ISBN numbers.  The implementation will always return valid ISBN numbers for english speaking books.  There is support
for other countries via overrides in the yml files but the values are not currently set.  Default values are constructed in the
Java so that the functionality doesn't break if we bring over the yml files from the stympy/faker project.

A test is included however to show that the values can be customized via yml if we choose to do so.

Fixes DiUS/java-faker#168

* Quiet down the logs for tests, there is too much logging