Clarify documentation of DateAndTime.birthday(..)
When using DateAndTime.birthday(..) it's not obvious that the generated
birthday depends on the current day. This leads to the wrong assumption
that calling the method with the same random seed creates the same
birthday with every call. However, the method produces different
results whether it's called in 2020 or 2021.

This commit adds clarification to the method's Javadoc.